backend/trackBot/requirements.txt
langchain
langchain-openai
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
pydantic>=1.8.2
httpx>=0.24.0
pydantic-settings
langgraph
python-dateutil
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
alembic>=1.7.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=41.0.0